refactor(web): resolve members-page review judgment calls
The escalation checkbox now carries the full selection count, the
org-removal preview warns when co-owners are selected, the typed
confirm phrase is localized, the bulk writes go through thin
hooks/mutations wrappers, and BulkActionsBar splits its result
builders, preview panel, and remove dialog into their own modules.
