+++ web/src/locales/en.json
@@ -3381,13 +3381,17 @@
       "removeModalTitle_other": "Remove {{count}} selected members from a classroom",
       "confirmRemoveBody_one": "This removes the selected member from the {{classroom}} classroom roster and its team. Their organization membership and assignment repositories are not affected.",
       "confirmRemoveBody_other": "This removes the selected members from the {{classroom}} classroom roster and its team. Their organization membership and assignment repositories are not affected.",
-      "alsoRemoveFromOrg": "Also remove from the {{org}} organization",
+      "alsoRemoveFromOrg_one": "Also remove the selected member from the {{org}} organization",
+      "alsoRemoveFromOrg_other": "Also remove all {{count}} selected members from the {{org}} organization",
       "confirmRemoveOrgTitle_one": "Remove {{count}} member from the {{org}} organization?",
       "confirmRemoveOrgTitle_other": "Remove {{count}} members from the {{org}} organization?",
       "confirmRemoveOrgBody_one": "The selected member will first be unenrolled from every classroom they belong to, then removed from the {{org}} organization. Their assignment repositories are not deleted.",
       "confirmRemoveOrgBody_other": "The selected members will first be unenrolled from every classroom they belong to, then removed from the {{org}} organization. Their assignment repositories are not deleted.",
       "orgRemoveOtherClassrooms_one": "{{count}} selected member also belongs to other classrooms and will be unenrolled from those too.",
       "orgRemoveOtherClassrooms_other": "{{count}} selected members also belong to other classrooms and will be unenrolled from those too.",
+      "orgRemoveOwners_one": "{{count}} of the selected members is an organization owner.",
+      "orgRemoveOwners_other": "{{count}} of the selected members are organization owners.",
+      "confirmPhrase": "confirm",
       "addModalTitle_one": "Add {{count}} selected member to a classroom",
       "addModalTitle_other": "Add {{count}} selected members to a classroom",
       "confirmAddBody_one": "This adds the selected member to the {{classroom}} classroom roster and its team, inviting them to the organization if they aren't a member yet.",
